requirements.txt
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.65.0
google-generativeai>=0.3.0
